Fix possible format specifier warning
Suppose someone specifies the lattice size at runtime and does
something like

<dimension name="j" type="integer" lattice = "numGridPoints" domain="(1, numGridPoints)" />

Then the "1" is treated as an int, and numGridPoints as a long. This
results in a compile time warning, as a printf statement expects all
its arguments to be long ints rather than ints.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2715 41b19578-b77a-4e3b-9682-c8570d237623

+++ xpdeint/XMDS2Parser.py
@@ -647,7 +647,7 @@ Use feature <validation kind="run-time"/> to allow for arbitrary code.""" % loca
             if ((%(latticeString)s) != ((%(maximumString)s) - (%(minimumString)s) + 1))
               _LOG(_ERROR_LOG_LEVEL, "ERROR: The lattice value of '%(latticeString)s'=%%li doesn't match with the domain "
                                      "'%(minimumString)s'=%%li to '%(maximumString)s'=%%li (%%li lattice points).\\n",
-                                     %(latticeString)s, %(minimumString)s, %(maximumString)s, ((%(maximumString)s) - (%(minimumString)s)+1));
+                                     long(%(latticeString)s), long(%(minimumString)s), long(%(maximumString)s), long((%(maximumString)s) - (%(minimumString)s)+1));
             """ % locals())
         else:
           # Only for 'long' dimensions
